+++ src/test/java/uk/gov/hmcts/reform/iacasenotificationsapi/domain/personalisation/appellant/letter/AppellantInternalCaseSubmitAppealOutOfTimeWithRemissionLetterPersonalisationTest.java
@@ -1,8 +1,9 @@
-package uk.gov.hmcts.reform.iacasenotificationsapi.domain.personalisation.appellant;
+package uk.gov.hmcts.reform.iacasenotificationsapi.domain.personalisation.appellant.letter;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.when;
+import static uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.AsylumCaseDefinition.AMOUNT_LEFT_TO_PAY;
 
 import java.util.Map;
 import java.util.Optional;
@@ -15,18 +16,17 @@ import org.mockito.junit.jupiter.MockitoSettings;
 import org.mockito.quality.Strictness;
 import uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.AsylumCase;
 import uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.AsylumCaseDefinition;
-import uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.Direction;
-import uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.DirectionTag;
 import uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.ccd.CaseDetails;
 import uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.ccd.callback.Callback;
 import uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.ccd.field.AddressUk;
-import uk.gov.hmcts.reform.iacasenotificationsapi.domain.service.DirectionFinder;
+import uk.gov.hmcts.reform.iacasenotificationsapi.domain.personalisation.appellant.letter.AppellantInternalCaseSubmitAppealOutOfTimeWithRemissionLetterPersonalisation;
 import uk.gov.hmcts.reform.iacasenotificationsapi.infrastructure.CustomerServicesProvider;
+import uk.gov.hmcts.reform.iacasenotificationsapi.infrastructure.SystemDateProvider;
+
 
 @ExtendWith(MockitoExtension.class)
 @MockitoSettings(strictness = Strictness.LENIENT)
-class AppellantInternalHomeOfficeDirectedToReviewAppealPersonalisationTest {
-
+class AppellantInternalCaseSubmitAppealOutOfTimeWithRemissionLetterPersonalisationTest {
     @Mock
     Callback<AsylumCase> callback;
     @Mock
@@ -36,11 +36,8 @@ class AppellantInternalHomeOfficeDirectedToReviewAppealPersonalisationTest {
     @Mock
     CustomerServicesProvider customerServicesProvider;
     @Mock
-    DirectionFinder directionFinder;
-    @Mock
-    Direction direction;
-    @Mock
     AddressUk appellantAddress;
+
     private Long ccdCaseId = 12345L;
     private String letterTemplateId = "someLetterTemplateId";
     private String appealReferenceNumber = "someAppealRefNumber";
@@ -48,22 +45,21 @@ class AppellantInternalHomeOfficeDirectedToReviewAppealPersonalisationTest {
     private String appellantGivenNames = "someAppellantGivenNames";
     private String appellantFamilyName = "someAppellantFamilyName";
     private String addressLine1 = "50";
-    private String expectedDirectionDueDate = "27 Aug 2024";
-    private String directionDueDate = "2024-08-27";
-    private String directionExplanation = "someExplanation";
     private String addressLine2 = "Building name";
     private String addressLine3 = "Street name";
     private String postCode = "XX1 2YY";
     private String postTown = "Town name";
     private String customerServicesTelephone = "555 555 555";
     private String customerServicesEmail = "example@example.com";
-    private AppellantInternalHomeOfficeDirectedToReviewAppealPersonalisation appellantInternalHomeOfficeDirectedToReviewAppealPersonalisation;
+    private final SystemDateProvider systemDateProvider = new SystemDateProvider();
+    private String amountLeftToPay = "4000";
+    private String amountLeftToPayInGbp = "40.00";
+    private int daysAfterSubmitAppeal = 10;
+
+    private AppellantInternalCaseSubmitAppealOutOfTimeWithRemissionLetterPersonalisation appellantInternalCaseSubmitAppealOutOfTimeWithRemissionLetterPersonalisation;
 
     @BeforeEach
     public void setup() {
-        when((direction.getDateDue())).thenReturn(directionDueDate);
-        when((direction.getExplanation())).thenReturn(directionExplanation);
-        when(directionFinder.findFirst(asylumCase, DirectionTag.RESPONDENT_REVIEW)).thenReturn(Optional.of(direction));
 
         when(callback.getCaseDetails()).thenReturn(caseDetails);
         when(caseDetails.getId()).thenReturn(ccdCaseId);
@@ -81,32 +77,34 @@ class AppellantInternalHomeOfficeDirectedToReviewAppealPersonalisationTest {
         when(appellantAddress.getPostCode()).thenReturn(Optional.of(postCode));
         when(appellantAddress.getPostTown()).thenReturn(Optional.of(postTown));
 
-        appellantInternalHomeOfficeDirectedToReviewAppealPersonalisation = new AppellantInternalHomeOfficeDirectedToReviewAppealPersonalisation(
+        appellantInternalCaseSubmitAppealOutOfTimeWithRemissionLetterPersonalisation = new AppellantInternalCaseSubmitAppealOutOfTimeWithRemissionLetterPersonalisation(
             letterTemplateId,
-            customerServicesProvider, directionFinder);
+            daysAfterSubmitAppeal,
+            customerServicesProvider,
+            systemDateProvider);
     }
 
     @Test
     void should_return_given_template_id() {
-        assertEquals(letterTemplateId, appellantInternalHomeOfficeDirectedToReviewAppealPersonalisation.getTemplateId());
+        assertEquals(letterTemplateId, appellantInternalCaseSubmitAppealOutOfTimeWithRemissionLetterPersonalisation.getTemplateId());
     }
 
     @Test
     void should_return_given_reference_id() {
-        assertEquals(ccdCaseId + "_INTERNAL_HO_REVIEW_THE_APPEAL_APPELLANT_LETTER",
-            appellantInternalHomeOfficeDirectedToReviewAppealPersonalisation.getReferenceId(ccdCaseId));
+        assertEquals(ccdCaseId + "_INTERNAL_SUBMIT_APPEAL_OUT_OF_TIME_WITH_REMISSION_APPELLANT_LETTER",
+            appellantInternalCaseSubmitAppealOutOfTimeWithRemissionLetterPersonalisation.getReferenceId(ccdCaseId));
     }
 
     @Test
     void should_return_address_in_correct_format() {
-        assertTrue(appellantInternalHomeOfficeDirectedToReviewAppealPersonalisation.getRecipientsList(asylumCase).contains("50_Buildingname_Streetname_Townname_XX12YY"));
+        assertTrue(appellantInternalCaseSubmitAppealOutOfTimeWithRemissionLetterPersonalisation.getRecipientsList(asylumCase).contains("50_Buildingname_Streetname_Townname_XX12YY"));
     }
 
     @Test
     void should_throw_exception_when_cannot_find_address_for_appellant() {
         when(asylumCase.read(AsylumCaseDefinition.APPELLANT_ADDRESS, AddressUk.class)).thenReturn(Optional.empty());
 
-        assertThatThrownBy(() -> appellantInternalHomeOfficeDirectedToReviewAppealPersonalisation.getRecipientsList(asylumCase))
+        assertThatThrownBy(() -> appellantInternalCaseSubmitAppealOutOfTimeWithRemissionLetterPersonalisation.getRecipientsList(asylumCase))
             .isExactlyInstanceOf(IllegalStateException.class)
             .hasMessage("appellantAddress is not present");
     }
@@ -115,7 +113,7 @@ class AppellantInternalHomeOfficeDirectedToReviewAppealPersonalisationTest {
     void should_throw_exception_on_personalisation_when_case_is_null() {
 
         assertThatThrownBy(
-            () -> appellantInternalHomeOfficeDirectedToReviewAppealPersonalisation.getPersonalisation((Callback<AsylumCase>) null))
+            () -> appellantInternalCaseSubmitAppealOutOfTimeWithRemissionLetterPersonalisation.getPersonalisation((Callback<AsylumCase>) null))
             .isExactlyInstanceOf(NullPointerException.class)
             .hasMessage("callback must not be null");
     }
@@ -123,9 +121,12 @@ class AppellantInternalHomeOfficeDirectedToReviewAppealPersonalisationTest {
     @Test
     void should_return_personalisation_when_all_information_given() {
 
-        Map<String, String> personalisation =
-            appellantInternalHomeOfficeDirectedToReviewAppealPersonalisation.getPersonalisation(callback);
+        when(asylumCase.read(AMOUNT_LEFT_TO_PAY, String.class)).thenReturn(Optional.of(amountLeftToPay));
 
+        Map<String, String> personalisation =
+            appellantInternalCaseSubmitAppealOutOfTimeWithRemissionLetterPersonalisation.getPersonalisation(callback);
+
+        assertEquals(systemDateProvider.dueDate(10), personalisation.get("tenDaysAfterSubmitDate"));
         assertEquals(appellantGivenNames, personalisation.get("appellantGivenNames"));
         assertEquals(appellantFamilyName, personalisation.get("appellantFamilyName"));
         assertEquals(appealReferenceNumber, personalisation.get("appealReferenceNumber"));
@@ -137,7 +138,6 @@ class AppellantInternalHomeOfficeDirectedToReviewAppealPersonalisationTest {
         assertEquals(postCode, personalisation.get("address_line_5"));
         assertEquals(customerServicesTelephone, customerServicesProvider.getCustomerServicesTelephone());
         assertEquals(customerServicesEmail, customerServicesProvider.getCustomerServicesEmail());
-        assertEquals(expectedDirectionDueDate, personalisation.get("directionDueDate"));
-
+        assertEquals(amountLeftToPayInGbp, personalisation.get("feeAmount"));
     }
 }
