+++ src/test/java/uk/gov/hmcts/reform/iacasenotificationsapi/domain/personalisation/appellant/letter/AppellantInternalHomeOfficeDirectedToReviewAppealPersonalisationTest.java
@@ -1,9 +1,8 @@
-package uk.gov.hmcts.reform.iacasenotificationsapi.domain.personalisation.appellant;
+package uk.gov.hmcts.reform.iacasenotificationsapi.domain.personalisation.appellant.letter;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.when;
-import static uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.AsylumCaseDefinition.FEE_AMOUNT_GBP;
 
 import java.util.Map;
 import java.util.Optional;
@@ -16,17 +15,19 @@ import org.mockito.junit.jupiter.MockitoSettings;
 import org.mockito.quality.Strictness;
 import uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.AsylumCase;
 import uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.AsylumCaseDefinition;
+import uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.Direction;
+import uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.DirectionTag;
 import uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.ccd.CaseDetails;
 import uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.ccd.callback.Callback;
 import uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.ccd.field.AddressUk;
-import uk.gov.hmcts.reform.iacasenotificationsapi.domain.personalisation.appellant.letter.AppellantInternalCaseSubmittedOnTimeWithFeePersonalisation;
+import uk.gov.hmcts.reform.iacasenotificationsapi.domain.personalisation.appellant.letter.AppellantInternalHomeOfficeDirectedToReviewAppealPersonalisation;
+import uk.gov.hmcts.reform.iacasenotificationsapi.domain.service.DirectionFinder;
 import uk.gov.hmcts.reform.iacasenotificationsapi.infrastructure.CustomerServicesProvider;
-import uk.gov.hmcts.reform.iacasenotificationsapi.infrastructure.SystemDateProvider;
-
 
 @ExtendWith(MockitoExtension.class)
 @MockitoSettings(strictness = Strictness.LENIENT)
-class AppellantInternalCaseSubmittedOnTimeWithFeePersonalisationTest {
+class AppellantInternalHomeOfficeDirectedToReviewAppealPersonalisationTest {
+
     @Mock
     Callback<AsylumCase> callback;
     @Mock
@@ -36,8 +37,11 @@ class AppellantInternalCaseSubmittedOnTimeWithFeePersonalisationTest {
     @Mock
     CustomerServicesProvider customerServicesProvider;
     @Mock
+    DirectionFinder directionFinder;
+    @Mock
+    Direction direction;
+    @Mock
     AddressUk appellantAddress;
-
     private Long ccdCaseId = 12345L;
     private String letterTemplateId = "someLetterTemplateId";
     private String appealReferenceNumber = "someAppealRefNumber";
@@ -45,20 +49,22 @@ class AppellantInternalCaseSubmittedOnTimeWithFeePersonalisationTest {
     private String appellantGivenNames = "someAppellantGivenNames";
     private String appellantFamilyName = "someAppellantFamilyName";
     private String addressLine1 = "50";
+    private String expectedDirectionDueDate = "27 Aug 2024";
+    private String directionDueDate = "2024-08-27";
+    private String directionExplanation = "someExplanation";
     private String addressLine2 = "Building name";
     private String addressLine3 = "Street name";
     private String postCode = "XX1 2YY";
     private String postTown = "Town name";
     private String customerServicesTelephone = "555 555 555";
     private String customerServicesEmail = "example@example.com";
-    private final SystemDateProvider systemDateProvider = new SystemDateProvider();
-    private int daysAfterSubmitAppeal = 14;
-    private String amountLeftToPayInGbp = "140.00";
-    private String feeAmountGbp = "14000";
-    private AppellantInternalCaseSubmittedOnTimeWithFeePersonalisation appellantInternalCaseSubmittedOnTimeWithFeePersonalisation;
+    private AppellantInternalHomeOfficeDirectedToReviewAppealPersonalisation appellantInternalHomeOfficeDirectedToReviewAppealPersonalisation;
 
     @BeforeEach
     public void setup() {
+        when((direction.getDateDue())).thenReturn(directionDueDate);
+        when((direction.getExplanation())).thenReturn(directionExplanation);
+        when(directionFinder.findFirst(asylumCase, DirectionTag.RESPONDENT_REVIEW)).thenReturn(Optional.of(direction));
 
         when(callback.getCaseDetails()).thenReturn(caseDetails);
         when(caseDetails.getId()).thenReturn(ccdCaseId);
@@ -68,7 +74,6 @@ class AppellantInternalCaseSubmittedOnTimeWithFeePersonalisationTest {
         when(asylumCase.read(AsylumCaseDefinition.APPELLANT_ADDRESS, AddressUk.class)).thenReturn(Optional.of(appellantAddress));
         when(asylumCase.read(AsylumCaseDefinition.APPEAL_REFERENCE_NUMBER, String.class)).thenReturn(Optional.of(appealReferenceNumber));
         when(asylumCase.read(AsylumCaseDefinition.HOME_OFFICE_REFERENCE_NUMBER, String.class)).thenReturn(Optional.of(homeOfficeRefNumber));
-        when(asylumCase.read(FEE_AMOUNT_GBP, String.class)).thenReturn(Optional.of(feeAmountGbp));
         when((customerServicesProvider.getCustomerServicesTelephone())).thenReturn(customerServicesTelephone);
         when((customerServicesProvider.getCustomerServicesEmail())).thenReturn(customerServicesEmail);
         when(appellantAddress.getAddressLine1()).thenReturn(Optional.of(addressLine1));
@@ -77,35 +82,32 @@ class AppellantInternalCaseSubmittedOnTimeWithFeePersonalisationTest {
         when(appellantAddress.getPostCode()).thenReturn(Optional.of(postCode));
         when(appellantAddress.getPostTown()).thenReturn(Optional.of(postTown));
 
-
-        appellantInternalCaseSubmittedOnTimeWithFeePersonalisation = new AppellantInternalCaseSubmittedOnTimeWithFeePersonalisation(
+        appellantInternalHomeOfficeDirectedToReviewAppealPersonalisation = new AppellantInternalHomeOfficeDirectedToReviewAppealPersonalisation(
             letterTemplateId,
-            daysAfterSubmitAppeal,
-            customerServicesProvider,
-            systemDateProvider);
+            customerServicesProvider, directionFinder);
     }
 
     @Test
     void should_return_given_template_id() {
-        assertEquals(letterTemplateId, appellantInternalCaseSubmittedOnTimeWithFeePersonalisation.getTemplateId());
+        assertEquals(letterTemplateId, appellantInternalHomeOfficeDirectedToReviewAppealPersonalisation.getTemplateId());
     }
 
     @Test
     void should_return_given_reference_id() {
-        assertEquals(ccdCaseId + "_INTERNAL_SUBMIT_APPEAL_WITH_FEE_APPELLANT_LETTER",
-            appellantInternalCaseSubmittedOnTimeWithFeePersonalisation.getReferenceId(ccdCaseId));
+        assertEquals(ccdCaseId + "_INTERNAL_HO_REVIEW_THE_APPEAL_APPELLANT_LETTER",
+            appellantInternalHomeOfficeDirectedToReviewAppealPersonalisation.getReferenceId(ccdCaseId));
     }
 
     @Test
     void should_return_address_in_correct_format() {
-        assertTrue(appellantInternalCaseSubmittedOnTimeWithFeePersonalisation.getRecipientsList(asylumCase).contains("50_Buildingname_Streetname_Townname_XX12YY"));
+        assertTrue(appellantInternalHomeOfficeDirectedToReviewAppealPersonalisation.getRecipientsList(asylumCase).contains("50_Buildingname_Streetname_Townname_XX12YY"));
     }
 
     @Test
     void should_throw_exception_when_cannot_find_address_for_appellant() {
         when(asylumCase.read(AsylumCaseDefinition.APPELLANT_ADDRESS, AddressUk.class)).thenReturn(Optional.empty());
 
-        assertThatThrownBy(() -> appellantInternalCaseSubmittedOnTimeWithFeePersonalisation.getRecipientsList(asylumCase))
+        assertThatThrownBy(() -> appellantInternalHomeOfficeDirectedToReviewAppealPersonalisation.getRecipientsList(asylumCase))
             .isExactlyInstanceOf(IllegalStateException.class)
             .hasMessage("appellantAddress is not present");
     }
@@ -114,7 +116,7 @@ class AppellantInternalCaseSubmittedOnTimeWithFeePersonalisationTest {
     void should_throw_exception_on_personalisation_when_case_is_null() {
 
         assertThatThrownBy(
-            () -> appellantInternalCaseSubmittedOnTimeWithFeePersonalisation.getPersonalisation((Callback<AsylumCase>) null))
+            () -> appellantInternalHomeOfficeDirectedToReviewAppealPersonalisation.getPersonalisation((Callback<AsylumCase>) null))
             .isExactlyInstanceOf(NullPointerException.class)
             .hasMessage("callback must not be null");
     }
@@ -122,12 +124,9 @@ class AppellantInternalCaseSubmittedOnTimeWithFeePersonalisationTest {
     @Test
     void should_return_personalisation_when_all_information_given() {
 
-        when(asylumCase.read(FEE_AMOUNT_GBP, String.class)).thenReturn(Optional.of(feeAmountGbp));;
-
         Map<String, String> personalisation =
-            appellantInternalCaseSubmittedOnTimeWithFeePersonalisation.getPersonalisation(callback);
+            appellantInternalHomeOfficeDirectedToReviewAppealPersonalisation.getPersonalisation(callback);
 
-        assertEquals(systemDateProvider.dueDate(14), personalisation.get("fourteenDaysAfterSubmitDate"));
         assertEquals(appellantGivenNames, personalisation.get("appellantGivenNames"));
         assertEquals(appellantFamilyName, personalisation.get("appellantFamilyName"));
         assertEquals(appealReferenceNumber, personalisation.get("appealReferenceNumber"));
@@ -139,9 +138,7 @@ class AppellantInternalCaseSubmittedOnTimeWithFeePersonalisationTest {
         assertEquals(postCode, personalisation.get("address_line_5"));
         assertEquals(customerServicesTelephone, customerServicesProvider.getCustomerServicesTelephone());
         assertEquals(customerServicesEmail, customerServicesProvider.getCustomerServicesEmail());
-        assertEquals(amountLeftToPayInGbp, personalisation.get("feeAmount"));
+        assertEquals(expectedDirectionDueDate, personalisation.get("directionDueDate"));
+
     }
-
-
-
 }
